requirements.txt
numpy~=1.26.2
matplotlib==3.8.0
python-dotenv~=1.0.0
mesa~=2.1.5